Refactor project structure and improve codebase
- Move source files from root to standard Maven structure (src/main/java and src/test/java).
- Remove redundant TestUtils.java.
- Refactor WaitHelper to return WebElement and added Javadocs.
- Enhance DriverManager with requireDriver() and null-safe getDriver().
- Improve BaseTest with @Optional platform parameter for easier local testing.
- Update TestListener with null-checks and improved screenshot naming.
- Ensure all tests use updated fluent wait APIs.
