+++ src/main/java/utils/WaitHelper.java
@@ -23,18 +23,23 @@ public class WaitHelper {
     return ConfigReader.getInstance().getExplicitWait();
   }
 
+  // -------------------------------------------------------------------------
+  // Visibility waits
+  // -------------------------------------------------------------------------
+
   /**
-   * Waits for an element to be visible using a custom timeout.
+   * Waits for a WebElement to be visible and returns it.
    *
    * @param driver The Appium driver.
    * @param element The element to wait for.
-   * @param timeoutInSeconds The custom timeout in seconds.
+   * @param timeoutInSeconds Custom timeout in seconds.
+   * @return The visible element.
    */
-  public static void waitForElementToBeVisible(
+  public static WebElement waitForVisibility(
       AppiumDriver driver, WebElement element, int timeoutInSeconds) {
     try {
       WebDriverWait wait = new WebDriverWait(driver, Duration.ofSeconds(timeoutInSeconds));
-      wait.until(ExpectedConditions.visibilityOf(element));
+      return wait.until(ExpectedConditions.visibilityOf(element));
     } catch (Exception e) {
       logger.error("Element not visible within {} seconds", timeoutInSeconds, e);
       throw e;
@@ -42,27 +47,29 @@ public class WaitHelper {
   }
 
   /**
-   * Waits for an element to be visible using the default timeout.
+   * Waits for a WebElement to be visible using the default timeout and returns it.
    *
    * @param driver The Appium driver.
    * @param element The element to wait for.
+   * @return The visible element.
    */
-  public static void waitForElementToBeVisible(AppiumDriver driver, WebElement element) {
-    waitForElementToBeVisible(driver, element, getDefaultTimeout());
+  public static WebElement waitForVisibility(AppiumDriver driver, WebElement element) {
+    return waitForVisibility(driver, element, getDefaultTimeout());
   }
 
   /**
-   * Waits for an element located by a locator to be visible.
+   * Waits for an element located by a locator to be visible and returns it.
    *
    * @param driver The Appium driver.
    * @param locator The locator of the element.
-   * @param timeoutInSeconds The custom timeout in seconds.
+   * @param timeoutInSeconds Custom timeout in seconds.
+   * @return The visible element.
    */
-  public static void waitForElementToBeVisible(
+  public static WebElement waitForVisibility(
       AppiumDriver driver, By locator, int timeoutInSeconds) {
     try {
       WebDriverWait wait = new WebDriverWait(driver, Duration.ofSeconds(timeoutInSeconds));
-      wait.until(ExpectedConditions.visibilityOfElementLocated(locator));
+      return wait.until(ExpectedConditions.visibilityOfElementLocated(locator));
     } catch (Exception e) {
       logger.error(
           "Element with locator {} not visible within {} seconds", locator, timeoutInSeconds, e);
@@ -71,27 +78,34 @@ public class WaitHelper {
   }
 
   /**
-   * Waits for an element located by a locator to be visible using the default timeout.
+   * Waits for an element located by a locator to be visible using the default timeout and returns
+   * it.
    *
    * @param driver The Appium driver.
    * @param locator The locator of the element.
+   * @return The visible element.
    */
-  public static void waitForElementToBeVisible(AppiumDriver driver, By locator) {
-    waitForElementToBeVisible(driver, locator, getDefaultTimeout());
+  public static WebElement waitForVisibility(AppiumDriver driver, By locator) {
+    return waitForVisibility(driver, locator, getDefaultTimeout());
   }
 
+  // -------------------------------------------------------------------------
+  // Clickability waits
+  // -------------------------------------------------------------------------
+
   /**
-   * Waits for an element to be clickable using a custom timeout.
+   * Waits for a WebElement to be clickable and returns it.
    *
    * @param driver The Appium driver.
    * @param element The element to wait for.
-   * @param timeoutInSeconds The custom timeout in seconds.
+   * @param timeoutInSeconds Custom timeout in seconds.
+   * @return The clickable element.
    */
-  public static void waitForElementToBeClickable(
+  public static WebElement waitForClickability(
       AppiumDriver driver, WebElement element, int timeoutInSeconds) {
     try {
       WebDriverWait wait = new WebDriverWait(driver, Duration.ofSeconds(timeoutInSeconds));
-      wait.until(ExpectedConditions.elementToBeClickable(element));
+      return wait.until(ExpectedConditions.elementToBeClickable(element));
     } catch (Exception e) {
       logger.error("Element not clickable within {} seconds", timeoutInSeconds, e);
       throw e;
@@ -99,27 +113,29 @@ public class WaitHelper {
   }
 
   /**
-   * Waits for an element to be clickable using the default timeout.
+   * Waits for a WebElement to be clickable using the default timeout and returns it.
    *
    * @param driver The Appium driver.
    * @param element The element to wait for.
+   * @return The clickable element.
    */
-  public static void waitForElementToBeClickable(AppiumDriver driver, WebElement element) {
-    waitForElementToBeClickable(driver, element, getDefaultTimeout());
+  public static WebElement waitForClickability(AppiumDriver driver, WebElement element) {
+    return waitForClickability(driver, element, getDefaultTimeout());
   }
 
   /**
-   * Waits for an element located by a locator to be clickable using a custom timeout.
+   * Waits for an element located by a locator to be clickable and returns it.
    *
    * @param driver The Appium driver.
    * @param locator The locator of the element.
-   * @param timeoutInSeconds The custom timeout in seconds.
+   * @param timeoutInSeconds Custom timeout in seconds.
+   * @return The clickable element.
    */
-  public static void waitForElementToBeClickable(
+  public static WebElement waitForClickability(
       AppiumDriver driver, By locator, int timeoutInSeconds) {
     try {
       WebDriverWait wait = new WebDriverWait(driver, Duration.ofSeconds(timeoutInSeconds));
-      wait.until(ExpectedConditions.elementToBeClickable(locator));
+      return wait.until(ExpectedConditions.elementToBeClickable(locator));
     } catch (Exception e) {
       logger.error(
           "Element with locator {} not clickable within {} seconds", locator, timeoutInSeconds, e);
@@ -128,21 +144,151 @@ public class WaitHelper {
   }
 
   /**
-   * Waits for an element located by a locator to be clickable using the default timeout.
+   * Waits for an element located by a locator to be clickable using the default timeout and returns
+   * it.
    *
    * @param driver The Appium driver.
    * @param locator The locator of the element.
+   * @return The clickable element.
    */
-  public static void waitForElementToBeClickable(AppiumDriver driver, By locator) {
-    waitForElementToBeClickable(driver, locator, getDefaultTimeout());
+  public static WebElement waitForClickability(AppiumDriver driver, By locator) {
+    return waitForClickability(driver, locator, getDefaultTimeout());
+  }
+
+  // -------------------------------------------------------------------------
+  // Presence waits
+  // -------------------------------------------------------------------------
+
+  /**
+   * Waits for an element located by a locator to be present in the DOM and returns it.
+   *
+   * @param driver The Appium driver.
+   * @param locator The locator of the element.
+   * @param timeoutInSeconds Custom timeout in seconds.
+   * @return The present element.
+   */
+  public static WebElement waitForPresence(AppiumDriver driver, By locator, int timeoutInSeconds) {
+    try {
+      WebDriverWait wait = new WebDriverWait(driver, Duration.ofSeconds(timeoutInSeconds));
+      return wait.until(ExpectedConditions.presenceOfElementLocated(locator));
+    } catch (Exception e) {
+      logger.error(
+          "Element with locator {} not present within {} seconds", locator, timeoutInSeconds, e);
+      throw e;
+    }
   }
 
   /**
-   * Waits for an element to be invisible using a custom timeout.
+   * Waits for an element located by a locator to be present in the DOM using the default timeout
+   * and returns it.
+   *
+   * @param driver The Appium driver.
+   * @param locator The locator of the element.
+   * @return The present element.
+   */
+  public static WebElement waitForPresence(AppiumDriver driver, By locator) {
+    return waitForPresence(driver, locator, getDefaultTimeout());
+  }
+
+  // -------------------------------------------------------------------------
+  // Legacy void API — kept for backward compatibility
+  // -------------------------------------------------------------------------
+
+  /**
+   * @deprecated Use {@link #waitForVisibility(AppiumDriver, WebElement, int)} instead.
+   */
+  @Deprecated
+  public static void waitForElementToBeVisible(
+      AppiumDriver driver, WebElement element, int timeoutInSeconds) {
+    waitForVisibility(driver, element, timeoutInSeconds);
+  }
+
+  /**
+   * @deprecated Use {@link #waitForVisibility(AppiumDriver, WebElement)} instead.
+   */
+  @Deprecated
+  public static void waitForElementToBeVisible(AppiumDriver driver, WebElement element) {
+    waitForVisibility(driver, element);
+  }
+
+  /**
+   * @deprecated Use {@link #waitForVisibility(AppiumDriver, By, int)} instead.
+   */
+  @Deprecated
+  public static void waitForElementToBeVisible(
+      AppiumDriver driver, By locator, int timeoutInSeconds) {
+    waitForVisibility(driver, locator, timeoutInSeconds);
+  }
+
+  /**
+   * @deprecated Use {@link #waitForVisibility(AppiumDriver, By)} instead.
+   */
+  @Deprecated
+  public static void waitForElementToBeVisible(AppiumDriver driver, By locator) {
+    waitForVisibility(driver, locator);
+  }
+
+  /**
+   * @deprecated Use {@link #waitForClickability(AppiumDriver, WebElement, int)} instead.
+   */
+  @Deprecated
+  public static void waitForElementToBeClickable(
+      AppiumDriver driver, WebElement element, int timeoutInSeconds) {
+    waitForClickability(driver, element, timeoutInSeconds);
+  }
+
+  /**
+   * @deprecated Use {@link #waitForClickability(AppiumDriver, WebElement)} instead.
+   */
+  @Deprecated
+  public static void waitForElementToBeClickable(AppiumDriver driver, WebElement element) {
+    waitForClickability(driver, element);
+  }
+
+  /**
+   * @deprecated Use {@link #waitForClickability(AppiumDriver, By, int)} instead.
+   */
+  @Deprecated
+  public static void waitForElementToBeClickable(
+      AppiumDriver driver, By locator, int timeoutInSeconds) {
+    waitForClickability(driver, locator, timeoutInSeconds);
+  }
+
+  /**
+   * @deprecated Use {@link #waitForClickability(AppiumDriver, By)} instead.
+   */
+  @Deprecated
+  public static void waitForElementToBeClickable(AppiumDriver driver, By locator) {
+    waitForClickability(driver, locator);
+  }
+
+  /**
+   * @deprecated Use {@link #waitForPresence(AppiumDriver, By, int)} instead.
+   */
+  @Deprecated
+  public static void waitForPresenceOfElement(
+      AppiumDriver driver, By locator, int timeoutInSeconds) {
+    waitForPresence(driver, locator, timeoutInSeconds);
+  }
+
+  /**
+   * @deprecated Use {@link #waitForPresence(AppiumDriver, By)} instead.
+   */
+  @Deprecated
+  public static void waitForPresenceOfElement(AppiumDriver driver, By locator) {
+    waitForPresence(driver, locator);
+  }
+
+  // -------------------------------------------------------------------------
+  // Invisibility waits
+  // -------------------------------------------------------------------------
+
+  /**
+   * Waits for an element to be invisible.
    *
    * @param driver The Appium driver.
    * @param element The element to wait for.
-   * @param timeoutInSeconds The custom timeout in seconds.
+   * @param timeoutInSeconds Custom timeout in seconds.
    */
   public static void waitForElementToBeInvisible(
       AppiumDriver driver, WebElement element, int timeoutInSeconds) {
@@ -166,41 +312,12 @@ public class WaitHelper {
   }
 
   /**
-   * Waits for the presence of an element using a custom timeout.
-   *
-   * @param driver The Appium driver.
-   * @param locator The locator of the element.
-   * @param timeoutInSeconds The custom timeout in seconds.
-   */
-  public static void waitForPresenceOfElement(
-      AppiumDriver driver, By locator, int timeoutInSeconds) {
-    try {
-      WebDriverWait wait = new WebDriverWait(driver, Duration.ofSeconds(timeoutInSeconds));
-      wait.until(ExpectedConditions.presenceOfElementLocated(locator));
-    } catch (Exception e) {
-      logger.error(
-          "Element with locator {} not present within {} seconds", locator, timeoutInSeconds, e);
-      throw e;
-    }
-  }
-
-  /**
-   * Waits for the presence of an element using the default timeout.
-   *
-   * @param driver The Appium driver.
-   * @param locator The locator of the element.
-   */
-  public static void waitForPresenceOfElement(AppiumDriver driver, By locator) {
-    waitForPresenceOfElement(driver, locator, getDefaultTimeout());
-  }
-
-  /**
    * Waits for an element to disappear from the DOM.
    *
    * @param driver The Appium driver.
    * @param locator The locator of the element.
-   * @param timeoutInSeconds The custom timeout in seconds.
-   * @return True if the element disappears, false otherwise.
+   * @param timeoutInSeconds Custom timeout in seconds.
+   * @return {@code true} if the element disappears within the timeout, {@code false} otherwise.
    */
   public static boolean waitForElementToDisappear(
       AppiumDriver driver, By locator, int timeoutInSeconds) {
@@ -214,12 +331,17 @@ public class WaitHelper {
     }
   }
 
+  // -------------------------------------------------------------------------
+  // Custom condition wait
+  // -------------------------------------------------------------------------
+
   /**
-   * A custom wait that uses a specific ExpectedCondition.
+   * Waits for a custom {@link ExpectedCondition} to be met.
    *
    * @param driver The Appium driver.
-   * @param timeoutInSeconds The custom timeout in seconds.
-   * @param condition The ExpectedCondition to wait for.
+   * @param timeoutInSeconds Custom timeout in seconds.
+   * @param condition The condition to wait for.
+   * @param <T> The return type of the condition.
    */
   public static <T> void customWait(
       AppiumDriver driver, int timeoutInSeconds, ExpectedCondition<T> condition) {
