+++ src/test/java/com/example/BaseTest.java
@@ -4,6 +4,7 @@ import io.appium.java_client.AppiumDriver;
 import java.net.MalformedURLException;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
+import org.testng.annotations.Optional;
 import org.testng.annotations.Parameters;
 import utils.DriverManager;
 
@@ -12,7 +13,7 @@ public class BaseTest {
 
   @BeforeMethod
   @Parameters({"platform"})
-  public void setUp(String platform) throws MalformedURLException {
+  public void setUp(@Optional("android") String platform) throws MalformedURLException {
     if (platform == null || platform.isEmpty()) {
       platform = "android";
     }
