+++ src/test/java/com/example/ExampleTest.java
@@ -10,12 +10,11 @@ public class ExampleTest extends BaseTest {
 
   @Test(description = "Verify app launches successfully")
   public void testAppLaunch() {
-    // Example: Verify an element is displayed
-    // Replace with your actual element locator
+    // Example: Verify an element is displayed after the app loads.
+    // Replace with your actual element locator.
     try {
-      WaitHelper.waitForPresenceOfElement(
-          driver, AppiumBy.id("com.example.app:id/main_screen"), 20);
-      WebElement element = driver.findElement(AppiumBy.id("com.example.app:id/main_screen"));
+      WebElement element =
+          WaitHelper.waitForPresence(driver, AppiumBy.id("com.example.app:id/main_screen"), 20);
       Assert.assertTrue(element.isDisplayed(), "App did not launch successfully");
     } catch (Exception e) {
       Assert.fail("Failed to launch app: " + e.getMessage());
@@ -25,16 +24,15 @@ public class ExampleTest extends BaseTest {
   @Test(description = "Example test with actions")
   public void testBasicInteraction() {
     try {
-      // Example: Find and click a button
-      WaitHelper.waitForElementToBeClickable(driver, AppiumBy.accessibilityId("buttonId"), 15);
-      WebElement button = driver.findElement(AppiumBy.accessibilityId("buttonId"));
+      // Wait for button to be clickable, then click.
+      WebElement button =
+          WaitHelper.waitForClickability(driver, AppiumBy.accessibilityId("buttonId"), 15);
       button.click();
 
-      // Verify result
-      WaitHelper.waitForPresenceOfElement(
-          driver, AppiumBy.xpath("//android.widget.TextView[@text='Expected Result']"), 15);
+      // Verify result element appears.
       WebElement resultText =
-          driver.findElement(AppiumBy.xpath("//android.widget.TextView[@text='Expected Result']"));
+          WaitHelper.waitForPresence(
+              driver, AppiumBy.xpath("//android.widget.TextView[@text='Expected Result']"), 15);
       Assert.assertTrue(resultText.isDisplayed(), "Expected result not displayed");
     } catch (Exception e) {
       Assert.fail("Test interaction failed: " + e.getMessage());
